Update 1.0.1
I just messed around with the input to see if GetAxisRaw would be better for multiple inputs but I think I will rather just stick with arrow keys as it seems much simpler. I might go and work on this tonight if I'm not doing homework and I'm not tired.